Many small requested changes
Added an opening hours edit section to the main admin page.
Changed 'House Pours" to "Mainstays" on the beer section.
Put the opening hours in an xml format.
Put opening hours in a separate include php file.
Put the last line of the introduction on in its own paragraph.
Changed "main" menu to "food".
Added in editing functionality for the menu.